Separated proposal copy service and added attachment related notifications
